requirements.txt
djangorestframework_simplejwt==4.8.0
Django==3.2.6
django_heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
whitenoise==5.3.0
django-cors-headers==3.8.0
distlib==0.3.2
django-filter==2.4.0
django-storages==1.11.1
django-js-asset==1.2.2
